1.ContentCatgory的实体数据类型 2.修改ContentCatgory的mapper里querAll的sql语句

+++ manager-interface/src/main/java/com/shengde/e3mall/service/ItemCatService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface ItemCatService {
 
-    List<EasyUITreeNode> getItemCatlist(long parentId);
+    List<EasyUITreeNode> getItemCatlist(int parentId);
 }
